rename otbucket pkg to ot and add package docs
